#42 refactor: 패키지 구조 변경
도메인 중심으로 어댑터와 포트 개념 추가
