+++ src/main/java/com/fourdays/core/common/exception/InvalidException.java
@@ -1,0 +1,8 @@
+package com.fourdays.core.common.exception;
+
+public class InvalidException extends RuntimeException {
+
+    public InvalidException(String message) {
+        super(message);
+    }
+}
