--- src/main/java/com/fourdays/core/controller/dto/request/ShortenUrlDto.java
@@ -1,11 +1,0 @@
-package com.fourdays.core.controller.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Pattern;
-
-public record ShortenUrlDto(
-        @NotNull(message = "invalid url")
-        @Pattern(message = "invalid url", regexp = "^(http(s)?://)([a-z0-9\\w]+\\.*)+\\.[a-z0-9]{2,}(/[a-zA-Z0-9]*)?$")
-        String url
-) {
-}
