--- src/main/java/com/fourdays/core/controller/dto/response/OriginalUrlDto.java
@@ -1,4 +1,0 @@
-package com.fourdays.core.controller.dto.response;
-
-public record OriginalUrlDto(String url) {
-}
