--- src/main/java/com/fourdays/core/controller/dto/response/UrlKeyDto.java
@@ -1,4 +1,0 @@
-package com.fourdays.core.controller.dto.response;
-
-public record UrlKeyDto(String urlKey) {
-}
