--- src/main/java/com/fourdays/core/domain/url/entity/exception/InvalidDomainException.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.entity.exception;
-
-public class InvalidDomainException extends InvalidException {
-
-    public InvalidDomainException(String message) {
-        super(message);
-    }
-}
