--- src/main/java/com/fourdays/core/domain/url/entity/exception/InvalidException.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.entity.exception;
-
-public class InvalidException extends RuntimeException {
-
-    public InvalidException(String message) {
-        super(message);
-    }
-}
