--- src/main/java/com/fourdays/core/domain/url/entity/exception/InvalidPathException.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.entity.exception;
-
-public class InvalidPathException extends InvalidException {
-
-    public InvalidPathException(String message) {
-        super(message);
-    }
-}
