--- src/main/java/com/fourdays/core/domain/url/entity/exception/InvalidPortException.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.entity.exception;
-
-public class InvalidPortException extends InvalidException {
-
-    public InvalidPortException(String message) {
-        super(message);
-    }
-}
