--- src/main/java/com/fourdays/core/domain/url/entity/exception/InvalidProtocolException.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.entity.exception;
-
-public class InvalidProtocolException extends InvalidException {
-
-    public InvalidProtocolException(String message) {
-        super(message);
-    }
-}
