--- src/main/java/com/fourdays/core/domain/url/entity/exception/InvalidUrlException.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.entity.exception;
-
-public class InvalidUrlException extends InvalidException {
-
-    public InvalidUrlException(String message) {
-        super(message);
-    }
-}
