--- src/main/java/com/fourdays/core/domain/url/entity/exception/InvalidUrlKeyException.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.entity.exception;
-
-public class InvalidUrlKeyException extends InvalidException {
-
-    public InvalidUrlKeyException(String message) {
-        super(message);
-    }
-}
