--- src/main/java/com/fourdays/core/domain/url/service/UrlService.java
@@ -1,8 +1,0 @@
-package com.fourdays.core.domain.url.service;
-
-public interface UrlService {
-
-    String shortenUrl(String url);
-
-    String findOriginalUrlByUrlKey(String key);
-}
