+++ src/main/java/com/fourdays/core/url/application/port/in/ProtocolService.java
@@ -1,0 +1,10 @@
+package com.fourdays.core.url.application.port.in;
+
+import com.fourdays.core.url.domain.Protocol;
+
+import java.util.Optional;
+
+public interface ProtocolService {
+
+    Optional<Protocol> findByName(String name);
+}
