+++ src/main/java/com/fourdays/core/url/application/port/in/UrlService.java
@@ -1,0 +1,8 @@
+package com.fourdays.core.url.application.port.in;
+
+public interface UrlService {
+
+    String shortenUrl(String url);
+
+    String findOriginalUrlByUrlKey(String key);
+}
