+++ src/main/java/com/fourdays/core/url/domain/Protocol.java
@@ -1,0 +1,33 @@
+package com.fourdays.core.url.domain;
+
+public class Protocol {
+
+    private final Integer seq;
+
+    private final String name;
+
+    public Protocol(Integer seq, String name) {
+        this.seq = seq;
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public boolean isHttp() {
+        return "HTTP".equalsIgnoreCase(this.name);
+    }
+
+    public boolean isHttps() {
+        return "HTTPS".equalsIgnoreCase(this.name);
+    }
+
+    @Override
+    public String toString() {
+        return "Protocol{" +
+                "seq=" + seq +
+                ", name='" + name + '\'' +
+                '}';
+    }
+}
