+++ src/main/java/com/fourdays/core/url/exception/InvalidDomainException.java
@@ -1,0 +1,10 @@
+package com.fourdays.core.url.exception;
+
+import com.fourdays.core.common.exception.InvalidException;
+
+public class InvalidDomainException extends InvalidException {
+
+    public InvalidDomainException(String message) {
+        super(message);
+    }
+}
