+++ src/main/java/com/fourdays/core/url/exception/InvalidPathException.java
@@ -1,0 +1,10 @@
+package com.fourdays.core.url.exception;
+
+import com.fourdays.core.common.exception.InvalidException;
+
+public class InvalidPathException extends InvalidException {
+
+    public InvalidPathException(String message) {
+        super(message);
+    }
+}
