+++ src/main/java/com/fourdays/core/url/exception/InvalidPortException.java
@@ -1,0 +1,10 @@
+package com.fourdays.core.url.exception;
+
+import com.fourdays.core.common.exception.InvalidException;
+
+public class InvalidPortException extends InvalidException {
+
+    public InvalidPortException(String message) {
+        super(message);
+    }
+}
