+++ src/main/java/com/fourdays/core/url/exception/InvalidProtocolException.java
@@ -1,0 +1,10 @@
+package com.fourdays.core.url.exception;
+
+import com.fourdays.core.common.exception.InvalidException;
+
+public class InvalidProtocolException extends InvalidException {
+
+    public InvalidProtocolException(String message) {
+        super(message);
+    }
+}
