+++ src/main/java/com/fourdays/core/url/exception/InvalidUrlException.java
@@ -1,0 +1,10 @@
+package com.fourdays.core.url.exception;
+
+import com.fourdays.core.common.exception.InvalidException;
+
+public class InvalidUrlException extends InvalidException {
+
+    public InvalidUrlException(String message) {
+        super(message);
+    }
+}
