+++ src/main/java/com/fourdays/core/url/exception/InvalidUrlKeyException.java
@@ -1,0 +1,10 @@
+package com.fourdays.core.url.exception;
+
+import com.fourdays.core.common.exception.InvalidException;
+
+public class InvalidUrlKeyException extends InvalidException {
+
+    public InvalidUrlKeyException(String message) {
+        super(message);
+    }
+}
