+++ src/main/java/com/fourdays/core/url/util/Base62Encoder.java
@@ -1,0 +1,14 @@
+package com.fourdays.core.url.util;
+
+public interface Base62Encoder {
+
+    char[] CODES = {
+            'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z',
+            'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z',
+            '0', '1', '2', '3', '4', '5', '6', '7', '8', '9'
+    };
+
+    int SIZE = CODES.length;
+
+    String encode();
+}
